refact: reafct commont code for async & sync myqlm backends
